refactor: move the rest of the campaign APIs to the common directory
