+++ frontend/src/test-utils/api/constants.ts
@@ -3,3 +3,6 @@ export const TWILIO_CREDENTIAL = 'test-twilio-cred'
 export const TELEGRAM_CREDENTIAL = 'test-telegram-cred'
 export const USER_EMAIL = 'testEmail@open.gov.sg'
 export const DEFAULT_FROM = 'Postman Test <donotreply@test.postman.gov.sg>'
+export const CSV_FILENAME = 'test_email_recipients.csv'
+export const PRESIGNED_URL =
+  'https://s3.ap-southeast-1.amazonaws.com/file-test.postman.gov.sg/test_params'
