+++ frontend/src/test-utils/api/interfaces.ts
@@ -5,6 +5,9 @@ export interface State {
   // Auth
   users: User[]
   curUserId: number
+
+  // Campaigns
+  campaigns: Campaign[]
 }
 
 interface User {
@@ -22,4 +25,29 @@ interface User {
 interface Credential {
   label: string
   type: string
+}
+
+export interface Template {
+  body: string
+  subject?: string
+  from?: string
+  params: string
+  reply_to?: string
+}
+
+export interface Campaign {
+  created_at: Date
+  demo_message_limit: number | null
+  halted: boolean
+  id: number
+  name: string
+  protect: boolean
+  type: string
+  valid: boolean
+
+  csv_filename: string | null
+  is_csv_processing: boolean
+  job_queue: any[]
+  num_recipients: number | null
+  template?: Template
 }
